% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCode.R
\docType{methods}
\name{genCode}
\alias{genCode}
\alias{genCode,SlotDirectiveList-method}
\alias{genCode,SlotDirective-method}
\alias{genCode,ContainsDirective-method}
\alias{genCode,SlotDefaultDirectiveList-method}
\alias{genCode,SlotDefaultDirectiveListe-method}
\alias{genCode,SlotDefaultDirective-method}
\alias{genCode,list-method}
\alias{genCode,SigTypeDirectiveList-method}
\alias{genCode,S4GenericRecipe-method}
\alias{genCode,S4MethodRecipe-method}
\alias{genCode,S4ClassRecipe-method}
\title{genCode}
\usage{
genCode(obj, ...)

\S4method{genCode}{SlotDirectiveList}(obj, ...)

\S4method{genCode}{SlotDirective}(obj, ...)

\S4method{genCode}{ContainsDirective}(obj, ...)

\S4method{genCode}{SlotDefaultDirectiveList}(obj, ...)

\S4method{genCode}{SlotDefaultDirective}(obj, ...)

\S4method{genCode}{list}(obj, ...)

\S4method{genCode}{SigTypeDirectiveList}(obj, ...)

\S4method{genCode}{S4GenericRecipe}(obj, ...)

\S4method{genCode}{S4MethodRecipe}(obj)

\S4method{genCode}{S4ClassRecipe}(obj)
}
\arguments{
\item{obj}{The object (generally a class definition, *Recipe object
or part thereof, to generate code from}

\item{...}{unused}
}
\value{
A character vector of R code, or list of multiple code blocks
}
\description{
genCode
}
