% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAccessors.R
\name{makeAccessors}
\alias{makeAccessors}
\title{Generate accessors for an existing class}
\usage{
makeAccessors(cls, funnames = structure(names(cls@slots), names =
  names(cls@slots)), skip = character(), libs = character(), ...)
}
\arguments{
\item{cls}{The class definition to generate acessors for}

\item{funnames}{named list of accessor name, slot name pairs. Defaults to
acessor functions named identically to the slots they access}

\item{skip}{Character vector of slots to NOT create accessors for}

\item{libs}{character vector of packages to load (via library) before
generating acessors}

\item{...}{Unused}
}
\value{
a ClassAccessors object for use with write_accessors
}
\description{
Generate accessors for an existing class
}
