% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file}
\alias{parse_file}
\title{parse_file}
\usage{
parse_file(file)
}
\arguments{
\item{file}{The file containing the instruction comments to parse}
}
\value{
an S4RecipeList contianing the recipes specified in the file.
}
\description{
Parse a file with roxygen2-like S4 specifications into an
S4RecipeList
}
