% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAccessors.R
\docType{methods}
\name{write_accessors}
\alias{write_accessors}
\alias{write_accessors,list-method}
\alias{write_accessors,ClassAccessors-method}
\alias{write_accessors,SlotMethods-method}
\title{write accessor functions to file}
\usage{
write_accessors(info, filename = "./R/accessors.R", append,
  splash = .splashtext)

\S4method{write_accessors}{list}(info, filename = "./R/accessors.R", append,
  splash = .splashtext)

\S4method{write_accessors}{ClassAccessors}(info, filename = "./R/accessors.R",
  append, splash = .splashtext)

\S4method{write_accessors}{SlotMethods}(info, filename = "./R/accessors.R",
  append, splash = .splashtext)
}
\arguments{
\item{info}{the object containing information about the accessors
to generate}

\item{filename}{The file to write the generated accessors into.}

\item{append}{Should the accessors be appended to \code{filename} or
replace the file entirely.}

\item{splash}{The splashtext to place at the top of the file}
}
\description{
write accessor functions to file
}
\note{
In future versions of S4Coffee, this API will be deprecated
after it is unified with the genCode/writeCode workflow, but it
remains separate for now.
}
