% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCode.R
\name{writeCode}
\alias{writeCode}
\title{writeCode}
\usage{
writeCode(obj, outfile, append = FALSE)
}
\arguments{
\item{obj}{The Recipe or list of Recipes to write code for}

\item{outfile}{The filename or connection to write to}

\item{append}{if outfile is a connection, should code be appended rather
than replacing existing code. Defaults to FALSE}
}
\value{
NULL (silently)
}
\description{
write generated code to a file
}
